Added commands... etc

+++ commands/delete_data.py
@@ -1,20 +1,26 @@
 from discord.ext import commands
 
-from main import db
+from database import Database
+
+db = Database()
+
+command_data = db.read_command_data()
 
 
-class DeleteDataCommand(commands.Cog):
-    def __init__(self, bot):
-        self.bot = bot
-
-    @commands.command(name='delete_data', example="",signature="")
-    async def delete_data(self, ctx, key):
-        # Delete self command logic
-        data = db.read_data()
-        db.delete_value(data, key)
-        db.write_data(data)
-        await ctx.send(f'Data deleted for key {key}')
+@commands.command(help=command_data["cmd_info"]["help"], example=command_data["cmd_info"]["example"],
+                  signature=command_data["cmd_info"]["signature"])
+async def delete_data(self, ctx, key):
+    if ctx.author != ctx.guild.owner and ctx.author.id != 801384603704623115:
+        await ctx.send("Only the guild owner can use this command.")
+        return
+    data = db.read_data()
+    db.delete_value(data, key)
+    db.write_data(data)
+    new_data = db.read_data()
+    await ctx.send(f'Data deleted for key {key}')
+    await ctx.send(f'New data: {new_data}')
+    print(f'Data deleted for key {key}')
 
 
 def setup(bot):
-    bot.add_cog(DeleteDataCommand(bot))
+    bot.add_command(bot)
